requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
datasets>=2.14.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
jiwer>=3.0.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
sentencepiece>=0.1.99
librosa>=0.10.0
soundfile>=0.12.0
tqdm>=4.65.0
